#31 Add writer-requested session username when deleting / change log(sysout->logger)

+++ src/main/java/com/practice/Skilltest/board/service/impl/BoardServiceImpl.java
@@ -3,8 +3,12 @@ package com.practice.Skilltest.board.service.impl;
 import com.practice.Skilltest.board.dao.BoardDao;
 import com.practice.Skilltest.board.dto.BoardDto;
 import com.practice.Skilltest.board.service.BoardService;
+import com.practice.Skilltest.user.role.UserRoles;
 import org.apache.ibatis.javassist.NotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Repository;
 
 import java.util.Collection;
@@ -13,6 +17,8 @@ import java.util.Map;
 
 @Repository
 public class BoardServiceImpl implements BoardService {
+
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     private final BoardDao boardDao;
 
@@ -68,22 +74,33 @@ public class BoardServiceImpl implements BoardService {
 
     //게시글 삭제
     @Override
-    public void deleteBoard(long id) {
+    public void deleteBoard(long id, String username, Collection<? extends GrantedAuthority> userAuthority) {
+
+        if(!checkValidModify(id,username,userAuthority)) { return; }
+
         boardDao.deleteBoard(id);
     }
 
 
 
-    //요청자가
+    //요청자의 세션 정보가 해당 게시글의 작성자와 동일자 요청//어드민의 요청인지 확인
     @Override
     public boolean checkValidModify(long id, String username, Collection<? extends GrantedAuthority> userAuthority) {
 
+        //요청이 해당 게시글의 게시자와 동일한 유저가 요청하였는지 확인
+        if(boardDao.getWriter(id).equals(username)){ return true; }
 
+        logger.info("삭제자가 게시자와 다른 요청");
+        //만약 아니라면 해당 요청이 어드민 요청인지 확인
+        if(userAuthority.contains(new SimpleGrantedAuthority(UserRoles.ADMIN.getValue()))) {
+            
+            logger.info("어드민 게시글 삭제요청");
+            return true;
+        }
 
-
+        logger.info("삭제자가 게시자와 다르고 어드민이 아닌 요청");
         return false;
     }
-
 
 }
